tests, codestyle, after review
