+++ app/src/androidTest/java/com/androiddev/timerbreathing/HiltExt.kt
@@ -1,0 +1,41 @@
+package com.androiddev.timerbreathing
+
+import android.content.ComponentName
+import android.content.Intent
+import android.os.Bundle
+import androidx.annotation.StyleRes
+import androidx.core.util.Preconditions
+import androidx.fragment.app.Fragment
+import androidx.test.core.app.ActivityScenario
+import androidx.test.core.app.ApplicationProvider
+import com.androiddev.timerbreathingtest.HiltTestActivity
+
+inline fun <reified T : Fragment> launchFragmentInHiltContainer(
+    fragmentArgs: Bundle? = null,
+    @StyleRes themeResId: Int = R.style.FragmentScenarioEmptyFragmentActivityTheme,
+    crossinline action: Fragment.() -> Unit = {}
+    ) {
+        val startActivityIntent = Intent.makeMainActivity(
+            ComponentName(
+                ApplicationProvider.getApplicationContext(),
+                HiltTestActivity::class.java
+            )
+        ).putExtra(
+            "androidx.fragment.app.testing.FragmentScenario.EmptyFragmentActivity.THEME_EXTRAS_BUNDLE_KEY",
+            themeResId
+        )
+
+        ActivityScenario.launch<HiltTestActivity>(startActivityIntent).onActivity { activity ->
+            val fragment: Fragment = activity.supportFragmentManager.fragmentFactory.instantiate(
+                Preconditions.checkNotNull(T::class.java.classLoader),
+                T::class.java.name
+            )
+            fragment.arguments = fragmentArgs
+            activity.supportFragmentManager
+                .beginTransaction()
+                .add(android.R.id.content, fragment, "")
+                .commitNow()
+
+            fragment.action()
+        }
+    }
