+++ app/src/debug/AndroidManifest.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.timerbreathing">
+    package="com.androiddev.timerbreathing">
 
     <application>
-        <activity android:name="com.example.timerbreathingtest.HiltTestActivity"
+        <activity android:name="com.androiddev.timerbreathingtest.HiltTestActivity"
             android:exported="false"/>
     </application>
