+++ app/src/debug/java/com/androiddev/timerbreathingtest/HiltTestActivity.kt
@@ -1,0 +1,7 @@
+package com.androiddev.timerbreathingtest
+
+import androidx.appcompat.app.AppCompatActivity
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class HiltTestActivity:AppCompatActivity()
