--- app/src/debug/java/com/example/timerbreathingtest/HiltTestActivity.kt
@@ -1,7 +1,0 @@
-package com.example.timerbreathingtest
-
-import androidx.appcompat.app.AppCompatActivity
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class HiltTestActivity:AppCompatActivity()
