+++ app/src/main/AndroidManifest.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.timerbreathing">
+    package="com.androiddev.timerbreathing">
 
     <application
-        android:name=".MyApplication"
+        android:name="com.androiddev.timerbreathing.MyApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.TimerBreathing">
         <activity
-            android:name=".presentation.activities.MainActivity"
+            android:name="com.androiddev.timerbreathing.presentation.activities.MainActivity"
             android:screenOrientation="portrait"
             android:exported="true">
             <intent-filter>
@@ -20,7 +20,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <service android:name=".exoplayer.MusicService"
+        <service android:name="com.androiddev.timerbreathing.exoplayer.MusicService"
             android:exported="false">
             <intent-filter>
                 <action android:name="android.media.browse.MediaBrowserService"/>
