+++ app/src/main/java/com/androiddev/timerbreathing/MyApplication.kt
@@ -1,0 +1,7 @@
+package com.androiddev.timerbreathing
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MyApplication:Application()
