+++ app/src/main/java/com/androiddev/timerbreathing/data/ExerciseParameters.kt
@@ -1,0 +1,9 @@
+package com.androiddev.timerbreathing.data
+
+data class ExerciseParameters(
+    var timeTraining:Long,
+    var timeBreath:Long,
+    var timeExhalation:Long,
+    var timeBreathDelay:Long,
+    var timeExhalationDelay:Long
+)
