+++ app/src/main/java/com/androiddev/timerbreathing/data/TimerState.kt
@@ -1,0 +1,7 @@
+package com.androiddev.timerbreathing.data
+
+sealed class TimerState(val dataTime: ExerciseParameters) {
+    class Started(dataTime: ExerciseParameters) : TimerState(dataTime)
+    class Stopped(dataTime:ExerciseParameters) : TimerState(dataTime)
+}
+
