+++ app/src/main/java/com/androiddev/timerbreathing/presentation/activities/MainActivity.kt
@@ -1,0 +1,14 @@
+package com.androiddev.timerbreathing.presentation.activities
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.androiddev.timerbreathing.R
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
