+++ app/src/main/java/com/androiddev/timerbreathing/presentation/fragments/infoscreen/schemescreen/BalanceFragment.kt
@@ -1,0 +1,22 @@
+package com.androiddev.timerbreathing.presentation.fragments.infoscreen.schemescreen
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.View
+import androidx.navigation.fragment.findNavController
+import com.androiddev.timerbreathing.R
+import com.androiddev.timerbreathing.databinding.FragmentBalanceBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class BalanceFragment : Fragment(R.layout.fragment_balance) {
+    private lateinit var binding: FragmentBalanceBinding
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        binding = FragmentBalanceBinding.bind(view)
+        binding.toolbar.setNavigationOnClickListener {findNavController().popBackStack() }
+        binding.relaxBtn.setOnClickListener {
+            findNavController().navigate(R.id.action_balanceFragment_to_breathFragment) }
+    }
+}
