+++ app/src/main/java/com/androiddev/timerbreathing/utils/Constants.kt
@@ -1,0 +1,7 @@
+package com.androiddev.timerbreathing.utils
+
+object Constants {
+    const val RESOURCE = "resource"
+    const val MEDIA_ROOT_ID = "root_id"
+    const val ARG_STARTED = "isStarted"
+}
