--- app/src/main/java/com/example/timerbreathing/MyApplication.kt
@@ -1,8 +1,0 @@
-package com.example.timerbreathing
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-import dagger.hilt.android.lifecycle.HiltViewModel
-
-@HiltAndroidApp
-class MyApplication:Application()
