--- app/src/main/java/com/example/timerbreathing/data/ExerciseParameters.kt
@@ -1,9 +1,0 @@
-package com.example.timerbreathing.data
-
-data class ExerciseParameters(
-    var timeTraining:Long,
-    var timeBreath:Long,
-    var timeExhalation:Long,
-    var timeBreathDelay:Long,
-    var timeExhalationDelay:Long
-)
