--- app/src/main/java/com/example/timerbreathing/data/TimerState.kt
@@ -1,7 +1,0 @@
-package com.example.timerbreathing.data
-
-sealed class TimerState(val dataTime: ExerciseParameters) {
-    class Started(dataTime: ExerciseParameters) : TimerState(dataTime)
-    class Stopped(dataTime:ExerciseParameters) : TimerState(dataTime)
-}
-
