--- app/src/main/java/com/example/timerbreathing/presentation/activities/MainActivity.kt
@@ -1,15 +1,0 @@
-package com.example.timerbreathing.presentation.activities
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.appcompat.widget.Toolbar
-import com.example.timerbreathing.R
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
