--- app/src/main/java/com/example/timerbreathing/utils/Constants.kt
@@ -1,7 +1,0 @@
-package com.example.timerbreathing.utils
-
-object Constants {
-    const val RESOURCE = "resource"
-    const val MEDIA_ROOT_ID = "root_id"
-    const val ARG_STARTED = "isStarted"
-}
